+++ src/args/arg_builder/option.rs
@@ -18,9 +18,9 @@ pub struct OptBuilder<'n, 'e> {
     pub blacklist: Option<Vec<&'e str>>,
     pub possible_vals: Option<Vec<&'e str>>,
     pub requires: Option<Vec<&'e str>>,
-    pub num_vals: Option<u8>,
-    pub min_vals: Option<u8>,
-    pub max_vals: Option<u8>,
+    pub num_vals: Option<u64>,
+    pub min_vals: Option<u64>,
+    pub max_vals: Option<u64>,
     pub val_names: Option<VecMap<&'e str>>,
     pub validator: Option<Rc<Fn(String) -> StdResult<(), String>>>,
     pub overrides: Option<Vec<&'e str>>,
@@ -83,7 +83,7 @@ impl<'n, 'e> OptBuilder<'n, 'e> {
         };
         if let Some(ref vec) = ob.val_names {
             if vec.len() > 1 {
-                ob.num_vals = Some(vec.len() as u8);
+                ob.num_vals = Some(vec.len() as u64);
             }
         }
         if let Some(ref p) = a.validator {
@@ -98,7 +98,7 @@ impl<'n, 'e> OptBuilder<'n, 'e> {
         ob
     }
 
-    pub fn write_help<W: io::Write>(&self, w: &mut W, tab: &str, longest: usize) -> io::Result<()> {
+    pub fn write_help<W: io::Write>(&self, w: &mut W, tab: &str, longest: usize, skip_pv: bool) -> io::Result<()> {
         debugln!("fn=write_help");
         // if it supports multiple we add '...' i.e. 3 to the name length
         try!(write!(w, "{}", tab));
@@ -145,7 +145,26 @@ impl<'n, 'e> OptBuilder<'n, 'e> {
             // 8 = tab + '-a, '.len()
             write_spaces!((longest + 8) - (self.to_string().len()), w);
         }
-        print_opt_help!(self, longest + 12, w);
+        if let Some(h) = self.help {
+            if h.contains("{n}") {
+                let mut hel = h.split("{n}");
+                if let Some(part) = hel.next() {
+                    try!(write!(w, "{}", part));
+                }
+                for part in hel {
+                    try!(write!(w, "\n"));
+                    write_spaces!(longest + 12, w);
+                    try!(write!(w, "{}", part));
+                }
+            } else {
+                try!(write!(w, "{}", h));
+            }
+            if !skip_pv {
+                if let Some(ref pv) = self.possible_vals {
+                    try!(write!(w, " [values: {}]", pv.join(", ")));
+                }
+            }
+        }
         write!(w, "\n")
     }
 }
@@ -192,13 +211,13 @@ impl<'n, 'e> AnyArg<'n, 'e> for OptBuilder<'n, 'e> {
     fn is_set(&self, s: ArgSettings) -> bool { self.settings.is_set(s) }
     fn has_switch(&self) -> bool { true }
     fn set(&mut self, s: ArgSettings) { self.settings.set(s) }
-    fn max_vals(&self) -> Option<u8> { self.max_vals }
-    fn num_vals(&self) -> Option<u8> { self.num_vals }
+    fn max_vals(&self) -> Option<u64> { self.max_vals }
+    fn num_vals(&self) -> Option<u64> { self.num_vals }
     fn possible_vals(&self) -> Option<&[&'e str]> { self.possible_vals.as_ref().map(|o| &o[..]) }
     fn validator(&self) -> Option<&Rc<Fn(String) -> StdResult<(), String>>> {
         self.validator.as_ref()
     }
-    fn min_vals(&self) -> Option<u8> { self.min_vals }
+    fn min_vals(&self) -> Option<u64> { self.min_vals }
     fn short(&self) -> Option<char> { self.short }
     fn long(&self) -> Option<&'e str> { self.long }
     fn val_delim(&self) -> Option<char> { self.val_delim }